(GH-727) Fix: Name is bad if response has params
Get-WebFileName returns a weird and incorrect file name when the
response url has parameters. Determine if the response url has
parameters prior to determining the file name. As a last resort, look
at the request url. If the request url has a filename at the end
(determined by whether it has an extension or not), then use it as long
as it doesn't contain a query as well.